requirements.txt
openai==1.57.0
transformers==4.38.2
faiss-cpu==1.8.0.post1
datasets==3.1.0
fastapi==0.115.5
sentence-transformers==2.5.1
pandas==1.4.1
kagglehub==0.2.9
rpyc~=6.0.1
pydantic~=2.10.3
numpy~=1.26.4
uvicorn==0.32.1

requests~=2.32.3